Revise terminology and comments
