+++ compiler+runtime/src/jank/jank/nrepl/server/core.jank
@@ -45,15 +45,18 @@
     (.is_connected client)))
 
 (defn read-message
-  "Read a complete message (TODO) from the client and decode it."
-  [client* remainder]
+  "Reads input from the boxed `jank.nrepl.server.native_client` CLIENT*
+  and previouls uncosumed input BUFFER, returning [msg buffer],
+  where msg is the parsed message or nil if incomplete and buffer
+  is the unconsumed input."
+  [client* buffer]
   (let [client  (cpp/unbox (:* jank.nrepl.server.native_client) client*)
-        data (.read_some client)
-        combined (str remainder data)
-        decoded (bencode/parse combined)]
-    (if decoded
-      decoded
-      [nil combined])))
+        chunk (.read_some client)
+        input (str buffer chunk)
+        result (bencode/parse input)]
+    (if result
+      result
+      [nil input])))
 
 (defn write-message
   "Encode the given message and write it to the client socket."
