+++ compiler+runtime/test/bash/nrepl-server/src/connection_harness.jank
@@ -1,0 +1,39 @@
+(ns connection-harness
+  (:include "connection_harness.hpp")
+  (:require [jank.nrepl.server.bencode :as b]))
+
+(def ^:dynamic *harness*
+  "The underlying C++ connection_harness instance.")
+(def ^:dynamic *driver*
+  "The boxed test `jank.nrepl.server.native_client` used to drive the system under test.")
+(def ^:dynamic *sut**
+  "The boxed system under test `jank.nrepl.server.native_client` representing the server side endpoint.")
+
+(defn- read-some!
+  "Read a data string from the boxed `jank.nrepl.server.native_client` CLIENT*. Blocks until data is available."
+  [client*]
+  (let [client (cpp/* (cpp/unbox (:* jank.nrepl.server.native_client) client*))]
+    (.read_some client)))
+
+(defn- write-some!
+  "Write a DATA string from the boxed `jank.nrepl.server.native_client` CLIENT*."
+  [client* data]
+  (let [client (cpp/* (cpp/unbox (:* jank.nrepl.server.native_client) client*))]
+    (.write_some client data)))
+
+(defmacro with-harness
+  "Creates a connection harness and binds *harness* to the underlying
+  connection_harness, *driver* to the test driver endpoint, and *sut*
+  to the system under test endpoint, automatically tearing down the
+  harness after BODY execution."
+  [& body]
+  `(let [harness*# (cpp/connection_harness.make)]
+     (binding [connection-harness/*harness* (cpp/box harness*#)
+               connection-harness/*driver* (cpp/box (~'.-driver harness*#))
+               connection-harness/*sut** (cpp/box (~'.-sut harness*#))]
+       (try
+         ~@body
+         ;; (catch cpp/std.exception e#
+         ;;   (println :error (~'.what e#)))
+         (finally
+           (~'.teardown (cpp/unbox (:* connection_harness) connection-harness/*harness*)))))))
