+++ compiler+runtime/test/bash/nrepl-server/src/connection_harness_test.jank
@@ -1,0 +1,84 @@
+(ns connection-harness-test
+  (:include "connection_harness.hpp")
+  (:require [clojure.test :refer [deftest is run-tests successful? testing]]
+            [jank.nrepl.server.bencode :as ben]
+            [jank.nrepl.server.core :as core]
+            [connection-harness :as sf :refer [with-harness *driver* read-some! *sut** *harness* write-some!]]))
+
+(deftest session-test
+  (let [harness* (cpp/connection_harness.make)
+        server* (.-server harness*)
+        driver* (.-driver harness*)
+        sut*    (.-sut harness*)]
+
+    (testing "connection initialization"
+      (is (cpp/!= cpp/nullptr server*))
+      (is (cpp/!= cpp/nullptr driver*))
+      (is (cpp/!= cpp/nullptr sut*)))
+
+    (testing "connected state"
+      (is (.is_connected (cpp/* driver*)))
+      (is (.is_connected (cpp/* sut*))))
+
+    (testing "data roundtrip"
+      (let [msg "hello"]
+        (.write_some (cpp/* driver*) msg)
+        (let [received (.read_some (cpp/* sut*))]
+          (is (= msg received)))))
+
+    (testing "server to client write"
+      (let [msg "world"]
+        (.write_some (cpp/* sut*) msg)
+        (let [received (.read_some (cpp/* driver*))]
+          (is (= msg received)))))
+
+    (testing "connection close"
+      (.close (cpp/* driver*))
+      (is (not (.is_connected (cpp/* driver*))))
+      (.close (cpp/* sut*))
+      (is (not (.is_connected (cpp/* sut*))))))
+  )
+
+(deftest read-write-fn-test
+  (let [harness* (cpp/connection_harness.make)
+        sut* (.-sut harness*)
+        driver* (.-driver harness*)]
+    (testing "write sends data"
+      (write-some! (cpp/box driver*) "hello")
+      (is (= "hello" (.read_some (cpp/* sut*)))))
+
+    (testing "read returns data"
+      (.write_some (cpp/* sut*) "world")
+      (is (= "world" (read-some! (cpp/box driver*)))))
+
+    (.teardown harness*)))
+
+(deftest with-harness-test
+  (with-harness
+    (testing "dynamic vars are bound"
+      (is (some? *harness*))
+      (is (some? *driver*))
+      (is (some? *sut**)))
+
+    (testing "harness get port"
+      (is (pos?  (-> (cpp/unbox (:* connection_harness) *harness*)
+                     .-server
+                     cpp/*
+                     .get_port))))
+
+    (testing "driver write, sut read with read-some!"
+      (write-some! *driver* "hello")
+      (is (= "hello" (read-some! *sut**))))
+
+    (testing "sut write, driver read"
+      (write-some! *sut** "response")
+      (is (= "response" (read-some! *driver*))))))
+
+
+(def this-nsym (ns-name *ns*))
+
+(defn -main []
+  (cpp/exit
+   (if (successful? (run-tests this-nsym))
+     0
+     1)))
