--- compiler+runtime/test/bash/nrepl-server/src/session_fixture_test.jank
@@ -1,84 +1,0 @@
-(ns session-fixture-test
-  (:include "session_fixture.hpp")
-  (:require [clojure.test :refer [deftest is run-tests successful? testing]]
-            [jank.nrepl.server.bencode :as ben]
-            [jank.nrepl.server.core :as core]
-            [session-fixture :as sf :refer [with-session *client* read-some! *server* *session* write-some!]]))
-
-(deftest session-test
-  (let [session* (cpp/session_fixture.make)
-        server*        (.-server session*)
-        test-client*   (.-driver session*)
-        server-client* (.-sut session*)]
-
-    (testing "connection initialization"
-      (is (cpp/!= cpp/nullptr server*))
-      (is (cpp/!= cpp/nullptr test-client*))
-      (is (cpp/!= cpp/nullptr server-client*)))
-
-    (testing "connected state"
-      (is (.is_connected (cpp/* test-client*)))
-      (is (.is_connected (cpp/* server-client*))))
-    
-    (testing "data roundtrip"
-      (let [msg "hello"]
-        (.write_some (cpp/* test-client*) msg)
-        (let [received (.read_some (cpp/* server-client*))]
-          (is (= msg received)))))
-
-    (testing "server to client write"
-      (let [msg "world"]
-        (.write_some (cpp/* server-client*) msg)
-        (let [received (.read_some (cpp/* test-client*))]
-          (is (= msg received)))))
-
-    (testing "connection close"
-      (.close (cpp/* test-client*))
-      (is (not (.is_connected (cpp/* test-client*))))
-      (.close (cpp/* server-client*))
-      (is (not (.is_connected (cpp/* server-client*))))))
-  )
-
-(deftest read-write-fn-test
-  (let [session* (cpp/session_fixture.make)
-        server-client* (.-sut session*)
-        test-client* (.-driver session*)]
-    (testing "write sends data"
-      (write-some! (cpp/box test-client*) "hello")
-      (is (= "hello" (.read_some (cpp/* server-client*)))))
-
-    (testing "read returns data"
-      (.write_some (cpp/* server-client*) "world")
-      (is (= "world" (read-some! (cpp/box test-client*)))))
-
-    (.cleanup session*)))
-
-(deftest with-session-test
-  (with-session
-    (testing "dynamic vars are bound"
-      (is (some? *session*))
-      (is (some? *client*))
-      (is (some? *server*)))
-
-    (testing "session get port"
-      (is (pos?  (-> (cpp/unbox (:* session_fixture) *session*)
-                     .-server
-                     cpp/*
-                     .get_port))))
-
-    (testing "client write, server read with read-some!"
-      (write-some! *client* "hello")
-      (is (= "hello" (read-some! *server*))))
-
-    (testing "server write, client read"
-      (write-some! *server* "response")
-      (is (= "response" (read-some! *client*))))))
-
-
-(def this-nsym (ns-name *ns*))
-
-(defn -main []
-  (cpp/exit
-   (if (successful? (run-tests this-nsym))
-     0
-     1)))
